+++ src/game/WingedTub.cpp
@@ -27,13 +27,13 @@ trippin::WingedTub::WingedTub(
     frame = tubFrameFirst;
 }
 
-void trippin::WingedTub::beforeTick(Uint32) {
+void trippin::WingedTub::beforeTick(int) {
     if (inactive && activation.shouldActivate(hitBox)) {
         inactive = false;
     }
 }
 
-void trippin::WingedTub::afterTick(Uint32 engineTicks) {
+void trippin::WingedTub::afterTick(int engineTicks) {
     if (inactive) {
         return;
     }
@@ -58,7 +58,7 @@ void trippin::WingedTub::afterTick(Uint32 engineTicks) {
         if (frame == FRAME_CLOUD_LAST) {
             expired = true;
         }
-    } else if (static_cast<int>(engineTicks) % sprite.getFramePeriodTicks() == 0) { // Case #3: Advance wings cycle
+    } else if (engineTicks % sprite.getFramePeriodTicks() == 0) { // Case #3: Advance wings cycle
         frame++;
         if (frame == tubFrameLast) {
             frame = tubFrameFirst;
